HOC for Badger Logic -> Re-use for custom Badger buttons
