+++ src/atoms/Small.js
@@ -1,0 +1,10 @@
+// @flow
+
+import styled from 'styled-components'
+
+
+const Small = styled.span`
+	font-size: 12px;
+	font-weight: 700;
+`;
+export default Small
