+++ src/atoms/Text.js
@@ -1,0 +1,11 @@
+// @flow
+
+import styled from 'styled-components'
+
+const Text = styled.p`
+  font-size: 16px;
+  margin: 0;
+  line-height: 1.1em;
+`
+
+export default Text;
