+++ src/atoms/index.js
@@ -1,3 +1,5 @@
 // @flow
 
 export { default as Button } from './Button';
+export { default as Small } from './Small';
+export { default as Text } from './Text'
