+++ src/components/BadgerBase/README.md
@@ -1,0 +1,12 @@
+# Badger Base (HOC)
+
+The base lifecycle and state methods for Badger functionality.
+Badger buttons and badges get wrapped in this Higher Order Component (HOC), which provides it with the key functions and state to deal with Badger Payments.
+ 
+
+ ### What it does
+ * Fetch and update prices
+ * handleClick functionality
+ * Mounting setup
+ * Unmounting cleanup
+ *
