+++ src/components/BadgerBase/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+import BadgerBase from './BadgerBase'
+
+export default BadgerBase;
